pyproject.toml
[tool.poetry]
name = "whoischeker"
version = "0.1.0"
description = "Telegram bot for monitoring WHOIS and DNS records changes"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.4.1"
sqlalchemy = "^2.0.28"
alembic = "^1.13.1"
pyyaml = "^6.0.1"
python-whois = "^0.8.0"
dnspython = "^2.6.1"
aiosqlite = "^0.20.0"
asyncio = "^3.4.3"
python-dotenv = "^1.0.1"
tabulate = "^0.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing" 